Add visibility:-fs-table-paginate-repeated-visible for paginated tables. (#32)
This allows to render some elements on the continuation of a repeated
table header on a new page. E.g. a span containing the word "continued".

For this to work the table must have -fs-table-paginate: paginate; set.

Using visibility the element is always correctly layouted, but only
painted after the table is broken into a new page.