For #73 - Building on work by @rototor - creates a Java2D renderer for paged output.
For now, I’ve copied accross the files I needed to alter to the java2d
sub-module to avoid breaking things. Eventually we can get rid of all
Java2D and Swing code from the core.

Currently the renderer and builder in the sub-module are for paged
output only but I think we can combine both paged and continuous in the
one builder/renderer.