Implement ObjectDrawer for <object> tags.
Implement a default factory for FSObjectDrawer, which allows to
register custom drawers.

The testrunner includes a simple example how to implement
a FSObjectDrawer.

Currently its only possible with the PdfBoxRendererBuilder to register
a FSObjectDrawerFactory. As soon as #77 is merged I will implement
the same for Java2D.

This is the final feature I need to migrate all my reports
to OpenHtmlToPdf, as we do some custom graphs in some reports.