Links in PDFs: Track link overlapping per page, not for all pages
Currently the area occupied by a link is tracked in a Set to avoid
putting two links above each other. No idea why this can happen normaly,
but it least it tries to avoid that. But this Set is not per page,
its global for all pages.

Guess whats happend in a report on the production system of a customer
of mine :)

They have tables spanning many pages, and magically all links which happen
to be exactly at the same position like a link on another page the link
does not get exported.

With this patch the Set is tracked per page.

Is this Set really needed? Why was it introduced in the first place?