We must close all font files when we are finished with the document. Otherwise this may leak file handles. Those file handles should be finalized() sometime in future by the GC, but this may take way to long before all file handles are exhausted.
It seems strange that you must subset a font to let its file handle close,
but thats the way PDFont is implemented. Usually you always use all the
fonts you load into your PDF document. But if you don't use all fonts,
the unused fonts leak if they are not subsetted.

We also must close all TrueTypeCollections we load.

And we don't load fonts provided by file instantly, but rather use a supplier
for the font. This more or less also workarounds the file handle leak, as
no PDFont is loaded till it is really needed. Before this change the PDFont
always has been loaded, even if it is not used. And of course was leaked if
the font was not subset.