For #208 - Fixed position absolute test with bottom unit set.
It turned out the problem was in the test rather than the main code. I was trying to test the bottom offset but with no static elements in the body the body had a zero height and so a bottom offset was not visible (being above the top page margin). I fix this by using a min-height for the body and now the bottom offset is working perfectly.