MAJOR BUG FIX - We were not loading font metrics when loading a font via the builder.
Instead we were silently using the font metrics of the fallback font. This would often produce acceptable results and thus was not discovered until now. The bad part of this fix is that all the text visual regression tests are now failing as they were affected by using the wrong font metrics and are therefore a pixel or so out.

IMPORTANT: This fix needs to be merged with the main branch regardless of whether this branch itself is merged in entirety.