PageSupplier instead of PDFCreationListener.onCreatePage (PR#334)
(See https://github.com/danfickle/openhtmltopdf/pull/334)

Based on the suggested by @danfickle, reverted the changes to
PDFCreationListener and PdfBoxRenderer, instead a PageSupplier can be
set via the PdfRendererBuilder.

The PageSupplier.requestPage allows the user to create a new page or
supply an existing one where the Html will be printed.

PdfRendererBuilder implements PageSupplier and simply creates a new page
and adds it to the PDDocument.