Fixed incorrect pageNumber arguments in shadow-page loop.
Added a brief explanation what "shadow-pages" are to the PageSupplier.
As @danfickle pointed out, the current-page index (i) should be used as
pageNumber argument for PageSupplier in PdfBoxRenderer when building the
shadow-pages and not the pdfPageIndex (which at this point has already
been upped by one).
Also first call to PageSupplier should (for the first page) should be
with pageNumber 0, not 1.