Fixes #445 - get attribute value when dealing with Jsoup document.
We were making the assumption that a namespace aware element (ie. one with a local name) would have namespace aware attributes (ie. also have a local name).

This is not the case for documents converted from Jsoup. It turns out that they use createElementNS for creating elements but the old setAttribute (not namespace aware) for creating attributes.