#477: Only use the scaling from the given transform
For whatever reason we only want the DPI scaling from the given transform, and
not any other property of the transform.

When adding a link to process later, the transform has at that moment also a translate in it.

Before 1ffd27175926d8d25294500f71cc39f651421fc8 the used transform was after the last page was
processed. At that moment the transform only had a scale component and no translate.