#482: Added Workaround for lines only consisting of soft-hypen.
This is a special case only for this single case: Line-Width is 0 and the
line only contains of a soft-hypen.  In this case we don't need to try to
break on the character level, as this will not work, but will lead to a
endless loop. So we pretend we already did try to break on the character
level. This ends the loop for this line.