Fixes #756 - Update Rhino to 1.7.13 to avoid CVE.
It should be noted Rhino is used only in the openhtmltopdf-examples module (actually contains the testing code) and this module is not distributed with a build.

Rhino is used to confirm that script can run in an SVG. We only have this test so as to confirm that disabling script works.