Improve error reporting for missing font
This modifies the PdfBox renderer to throw a new exception, FontNotFoundException, in the case of a missing font instead of trying to access an empty font list and throw IndexOutOfBoundsException. A check is added in PdfBoxFastOutputDevice.setFont(FSFont) to assert the font description isn't null. The methods AbstractOutputDevice.drawText(RenderingContext, InlineText) and ListItemPainter.drawText(RenderingContext, BlockBox, IdentValue) are modified to set the font specification before the font itself to make it available in the PdfBoxFastOutputDevice.setFont(FSFont) call.