Adding autorotation for JPEG
This change adds automatic image rotation for JPEG images with orientation EXIF data.

Almost all cell phone cameras store the image data in an un-rotated state, and write EXIF data to indicate "which side should be up".  All modern browsers understand these EXIF tags and automatically orient the image correctly.  Thus, when the pictures are viewed in a web page they appear correct.  When the page is output to PDF, however, the image is rotated.

This change modifies the pdfboximage class to read the EXIF data (if it exists) and rotate the image if necessary.  This results in a "rotate on load" behavior (i.e. before the resource cache), so the rotation only needs to be done once, even if the image appears multiple times in the document.